Salute (#83)
* Added salute png.

* OMGGGG ADDED SALUTEEEEE!

* Fixed all style imports.

* moved stylederror to common styles.
