+++ components/displayMessages/displayMessages.styles.tsx
@@ -1,3 +1,4 @@
+import { Divider } from '@mui/material'
 import { styled } from '@mui/material/styles'
 
 const StyledFinalMessagesContainer = styled('div')({
@@ -5,4 +6,9 @@ const StyledFinalMessagesContainer = styled('div')({
   marginBottom: 50
 })
 
-export { StyledFinalMessagesContainer }
+const StyledDivider = styled(Divider)({
+  marginTop: 50,
+  marginBottom: 50
+})
+
+export { StyledFinalMessagesContainer, StyledDivider }
