+++ styles/icons.ts
@@ -2,6 +2,10 @@ const icons = {
   HBP_LOGO: {
     image: '/assets/hbplogo.png',
     altText: 'HackBeanpot logo'
+  },
+  Salute: {
+    image: '/assets/salute.png',
+    altText: 'Salute'
   }
 }
